+++ tests/telescope/test_phosimCommu.py
@@ -27,11 +27,11 @@ from lsst.ts.wep.Utility import FilterType
 
 from lsst.ts.phosim.telescope.PhosimCommu import PhosimCommu
 
-from lsst.ts.phosim.Utility import getModulePath, SurfaceType
+from lsst.ts.phosim.utils.Utility import getModulePath, SurfaceType
 
 
 class TestPhosimCommu(unittest.TestCase):
-    """ Test the PhosimCommu class."""
+    """Test the PhosimCommu class."""
 
     def setUp(self):