+++ tests/utils/test_utility.py
@@ -25,7 +25,7 @@ import unittest
 from lsst.obs.lsst import LsstComCam, LsstCam
 from lsst.afw import cameraGeom
 
-from lsst.ts.phosim.Utility import (
+from lsst.ts.phosim.utils.Utility import (
     opt2ZemaxCoorTrans,
     zemax2optCoorTrans,
     mapSurfNameToEnum,
@@ -127,7 +127,11 @@ class TestUtility(unittest.TestCase):
         self.assertIsInstance(lsstComCam, cameraGeom.Camera)
         self.assertEqual(lsstComCam.getName(), LsstComCam.getCamera().getName())
 
-        lsstCam = getCamera("lsstfam")
+        lsstFamCam = getCamera("lsstfam")
+        self.assertIsInstance(lsstFamCam, cameraGeom.Camera)
+        self.assertEqual(lsstFamCam.getName(), LsstCam.getCamera().getName())
+
+        lsstCam = getCamera("lsst")
         self.assertIsInstance(lsstCam, cameraGeom.Camera)
         self.assertEqual(lsstCam.getName(), LsstCam.getCamera().getName())